# Improve speed of copying