@sewenthy 2024-06-30
Youtube removed my account for a bit for using this but on appeal they gave it back so I would say, use this at your own risk 😬 I'm closing this PR.